requirements.txt
et-xmlfile==1.0.1
lxml==4.6.3
openpyxl==3.0.7
Pillow==8.2.0
python-pptx==0.6.18
pytube==10.7.2
selenium==3.141.0
urllib3==1.26.4
XlsxWriter==1.3.9
